refactor: [US-048a] - Move dashboard/public-profile views into core/views/dashboard.py
